Feat[#15] - Participant Persistence & Query 계층 추가
