+++ backend/src/participant/participant.module.ts
@@ -1,23 +1,19 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { ParticipantService } from './participant.service';
 import { ParticipantController } from './participant.controller';
-import { MongooseModule } from '@nestjs/mongoose';
-import { ParticipantRepository } from './participant.repository';
-import { Participant, ParticipantSchema } from './schema/participant.schema';
 import { GroupBuyingModule } from '../group-buying/group-buying.module';
 import { GroupBuyingAccessGuard } from '../group-buying/guard/group-buying-access.guard';
 import { CommonModule } from '../common/common.module';
-import { GroupBuying, GroupBuyingSchema } from '../group-buying/schema/group-buying.schema';
 import { UserModule } from '../user/user.module';
 import { AuthModule } from '../auth/auth.module';
 import { WebPushModule } from '../web-push/web-push.module';
+import { ParticipantPersistenceModule } from './persistence/participant-persistence.module';
+import { ParticipantQueryModule } from './query/participant-query.module';
 
 @Module({
   imports: [
-    MongooseModule.forFeature([
-      { name: Participant.name, schema: ParticipantSchema },
-      { name: GroupBuying.name, schema: GroupBuyingSchema },
-    ]),
+    ParticipantPersistenceModule,
+    ParticipantQueryModule,
     forwardRef(() => GroupBuyingModule),
     CommonModule,
     UserModule,
@@ -25,7 +21,7 @@ import { WebPushModule } from '../web-push/web-push.module';
     WebPushModule,
   ],
   controllers: [ParticipantController],
-  providers: [ParticipantService, ParticipantRepository, GroupBuyingAccessGuard],
+  providers: [ParticipantService, GroupBuyingAccessGuard],
   exports: [ParticipantService, GroupBuyingAccessGuard],
 })
 export class ParticipantModule {}
